New fields, buttons & styles
Added the following values with their corresponding formulas:
-Height
-Excircle radius
-Incircle radius
These fields are hidden by default, but can be revealed with the "More Fields" button.

Another thing I updated is the navigation bar. With the 'navbar.css' file, I've given the buttons a design with a green theme and transitions for hovering and clicking. I've also made them appear on every page.